move calcite btn function to script
