+++ eds/scripts/scripts.js
@@ -13,7 +13,12 @@ import {
   loadScript,
 } from './aem.js';
 
-import { div, iframe, domEl } from './dom-helpers.js';
+import { 
+  div,
+  iframe,
+  domEl,
+  calciteButton,
+ } from './dom-helpers.js';
 
 const LCP_BLOCKS = ['header']; // add your LCP blocks to the list
 
@@ -149,6 +154,45 @@ function decorateMode(element) {
   });
 }
 
+/**
+* convert link button to calcite button
+* @param {element} block The block element
+*/
+function decorateToCalcite(main) {
+  main.querySelectorAll('p.button-container').forEach((buttonContainer) => {
+    // create a new div to replace the buttonContainer
+    const newDiv = div({ class: 'button-container' });
+    while (buttonContainer.firstChild) {
+    // if firstchild is <a> then construct calciteButton
+      if (buttonContainer.firstChild.tagName === 'A') {
+        newDiv.appendChild(calciteButton({
+          appearance: 'solid',
+          color: 'blue',
+          kind: 'brand',
+          scale: 'm',
+          href: buttonContainer.firstChild.href,
+          alignment: 'center',
+          type: 'button',
+          width: 'auto',
+        }, buttonContainer.firstChild.textContent));
+      } else if (buttonContainer.firstChild.tagName === 'EM') {
+        newDiv.appendChild(calciteButton({
+          appearance: 'outline',
+          color: 'blue',
+          kind: 'brand',
+          scale: 'm',
+          href: buttonContainer.firstChild.firstChild.href,
+          alignment: 'center',
+          type: 'button',
+          width: 'auto',
+        }, buttonContainer.firstChild.textContent));
+      }
+      buttonContainer.removeChild(buttonContainer.firstChild);
+    }
+    buttonContainer.replaceWith(newDiv);
+  });
+}
+
 export function decorateBlockMode(block) {
   decorateMode(block);
 }
@@ -191,6 +235,7 @@ async function loadEager(doc) {
   const main = doc.querySelector('main');
   if (main) {
     decorateMain(main);
+    decorateToCalcite(main);
     document.body.classList.add('appear');
     await waitForLCP(LCP_BLOCKS);
   }
